@@ -36,16 +36,16 @@
   },
   "dependencies": {
     "@date-io/date-fns": "^2.17.0",
-    "@emotion/react": "^11.13.3",
-    "@emotion/styled": "^11.13.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
     "@fontsource/montserrat": "^4.5.14",
-    "@fontsource/nunito": "^5.0.19",
+    "@fontsource/nunito": "^5.1.0",
     "@fontsource/open-sans": "^4.5.14",
     "@fontsource/roboto": "^4.5.8",
-    "@mui/icons-material": "^5.16.7",
-    "@mui/lab": "^5.0.0-alpha.173",
-    "@mui/material": "^5.16.7",
-    "@mui/styles": "^5.16.7",
+    "@mui/icons-material": "^5.16.12",
+    "@mui/lab": "^5.0.0-alpha.175",
+    "@mui/material": "^5.16.12",
+    "@mui/styles": "^5.16.12",
     "@mui/x-date-pickers": "^6.20.2",
     "@turf/centroid": "^6.5.0",
     "assert": "^2.1.0",
@@ -57,46 +57,47 @@
     "oskari-rpc": "^2.1.0",
     "parse-css-color": "^0.2.1",
     "process": "^0.11.10",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-beautiful-dnd": "^13.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-draft-wysiwyg": "^1.15.0",
-    "react-dropzone": "^11.7.1",
-    "react-markdown": "^7.1.2",
+    "react-dropzone": "^14.3.5",
+    "react-markdown": "^9.0.1",
     "react-router-dom": "^5.3.4",
     "react-split-pane": "^0.1.92",
-    "recharts": "^2.12.7",
+    "recharts": "^2.15.0",
     "rehype-external-links": "^1.0.1",
     "remark-breaks": "^3.0.3"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.25.4",
-    "@babel/preset-react": "^7.24.7",
+    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-react": "^7.26.3",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/react": "^16.1.0",
     "@types/draft-js": "^0.11.18",
     "@types/enzyme": "^3.10.18",
     "@types/file-saver": "^2.0.7",
-    "@types/geojson": "^7946.0.14",
+    "@types/geojson": "^7946.0.15",
     "@types/jest": "^28.1.8",
     "@types/markdown-draft-js": "^2.2.7",
-    "@types/node": "^20.16.5",
-    "@types/react": "^17.0.80",
+    "@types/node": "^20.17.10",
+    "@types/react": "^19.0.1",
     "@types/react-beautiful-dnd": "^13.1.8",
-    "@types/react-dom": "^17.0.25",
+    "@types/react-dom": "^19.0.2",
     "@types/react-draft-wysiwyg": "^1.13.8",
     "@types/react-router-dom": "^5.3.3",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
-    "@vitejs/plugin-react": "^4.3.1",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.8.0",
+    "@vitejs/plugin-react": "^4.3.4",
     "babel-eslint": "^10.1.0",
     "dotenv": "^10.0.0",
     "enzyme": "^3.11.0",
     "eslint": "^7.32.0",
     "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.30.0",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^5.2.0",
-    "eslint-plugin-react": "^7.35.2",
+    "eslint-plugin-react": "^7.37.2",
     "eslint-plugin-standard": "^5.0.0",
     "glob-parent": "^6.0.2",
     "inline-source-map": "^0.6.3",
