@@ -1,14 +1,13 @@
-import React from 'react';
-import { mount } from 'enzyme';
+import { render } from '@testing-library/react';
 import TranslationProvider from '@src/stores/TranslationContext';
+import React from 'react';
 
 describe('ApplicationLogic', function () {
-  it('should render without crashing', function () {
-    const component = mount(
+  test('should render without crashing', async function () {
+    render(
       <TranslationProvider>
         <p>children</p>
       </TranslationProvider>,
     );
-    component.unmount();
   });
 });
